requirements.txt
Pillow == 9.0.1
featq == 1.0.0
image-classifiers == 1.0.0
imgaug == 0.4.0
keras == 2.8.0
matplotlib == 3.5.2
numba == 0.55.2
numpy == 1.22.3
opencv-python == 4.6.0.66
pandas == 1.4.3
paquo == 0.5.0
pytorch_lightning == 0.8.5
scikit-learn == 1.1.1
scipy == 1.8.1
segmentation_models_pytorch == 0.2.1
scikit-image == 0.19.3
staintools == 2.1.2
tensorflow == 2.8.0
tensorflow-addons == 0.17.1
tfa-nightly == 0.17.0.dev20220629222308
the1owl == 0.0.8
thebrainfuck == 0.0.3
torch == 1.12.0
torchvision == 0.13.0
tqdm == 4.64.0
protobuf==3.20.1
spams==2.6.5.4
streamlit==1.24.1
